Finish the inicital functions
Add all the functions from the document of the investigation.

First we need to select the size and the number of threads and then click initialize and then click in go to see the matrix response.

When the result is print the time that threads using to have the response is print in a textBox.